% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_factors.R
\name{est_DESeq2_size_factors}
\alias{est_DESeq2_size_factors}
\title{Function to Calculate DESeq2 Size Factors}
\usage{
est_DESeq2_size_factors(counts)
}
\arguments{
\item{counts}{A (G x N) numeric matrix or data frame of RNA-seq counts, with genes in rows and samples in columns. G = number of genes.  N = number of samples.}
}
\value{
A vector with the size factors for each of the N samples
}
\description{
Estimates the size factors from DESeq2 to be used as offsets in NBGLMM model fits
}
\examples{
data("simdata")
counts <- simdata$counts

##Calculate size factors
size.factors <- est_DESeq2_size_factors(counts)



}
