% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmseq-pkg.R
\docType{package}
\name{mcmseq-pkg}
\alias{mcmseq-pkg}
\title{mcmseq: Bayesian Hierarchical Negative Binomial Mixed Models for RNA-Seq}
\description{
Fitting utilities for estimating negative binomial generalized linear mixed models for use
with RNA-Seq data from studies with clustered/longitudinal study designs
}
