% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_check.R
\name{mcmseq.convergence}
\alias{mcmseq.convergence}
\title{Function to Check for Convergence Issues in the MCMC Chains}
\usage{
mcmseq.convergence(mcmseqModel, geweke.p = 0.05,
  prop.accepts.betas = 0.1, prop.accepts.alphas = 0.1)
}
\arguments{
\item{mcmseqModel}{A model fit with the mcmseq.fit function.}

\item{geweke.p}{The p-value for the Geweke statistic that should be used to determine convergence.  By default Geweke p-values less than 0.05 for any parameter are considered convergence failures.}

\item{prop.accepts.betas}{The acceptance rate for regression coefficients below which the model is considered a convergence failure. Default is 0.1.}

\item{prop.accepts.alphas}{The acceptance rate for dispersion parameters below which the model is considered a convergence failure. Default is 0.1.}
}
\value{
A data frame including Geweke statistic and acceptance rates for genes that failed to converge.
}
\description{
Checks Geweke statistics and acceptance rates for model parameters to determine convergence issues.
}
\examples{
data("simdata")
metadata <- simdata$metadata

##Only including 10 genes in the counts matrix
counts <- simdata$counts[1:10,]

##Create the fixed effects model formula
f <- ~ group*time

##Fit model
fit.default <- mcmseq.fit(counts=counts,
                         fixed_effects = f,
                         sample_data = metadata,
                         random_intercept = 'ids',
                         gene_names = paste0('gene_', seq(1,10,1)),
                         n_it = 1000,
                         prop_burn_in = 0.1)

##Check convergence
fit.converge<-mcmseq.convergence(fit.default)


}
