% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbglm_mcmc_fp2}
\alias{nbglm_mcmc_fp2}
\title{Negative Binomial GLM MCMC WLS (full parallel chians)}
\usage{
nbglm_mcmc_fp2(counts, design_mat, prior_sd_betas, prior_sd_rs,
  prior_mean_log_rs, n_it, rw_sd_rs, log_offset, starting_betas,
  grain_size = 1L, return_summary = TRUE, burn_in_prop = 0.1, VIF = 1)
}
\arguments{
\item{counts}{a matrix of counts}

\item{design_mat}{design matrix for mean response}

\item{prior_sd_betas}{prior std. dev. for regression coefficients}

\item{prior_sd_rs}{prior std. dev for dispersion parameters}

\item{prior_mean_log_rs}{vector of prior means for dispersion parameters}

\item{n_it}{number of iterations to run MCMC}

\item{rw_sd_rs}{random wal std. dev. for proposing dispersion values}

\item{log_offset}{vector of offsets on log scale}

\item{grain_size}{minimum size of parallel jobs, defaults to 1, can ignore for now}

\item{return_summary}{return only a summary (point estimates and p-values where appropriate)}
}
\value{
Returns a list with a cube of regression parameters, and a matrix of dispersion values
}
\description{
Run an MCMC for the Negative Binomial mixed model (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
