% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbglm_mcmc_rw}
\alias{nbglm_mcmc_rw}
\title{Negative Binomial GLMM MCMC Random Walk (full parallel chians)}
\usage{
nbglm_mcmc_rw(counts, design_mat, contrast_mat, prior_sd_betas,
  rw_sd_betas, prior_sd_rs, prior_mean_log_rs, n_it, rw_sd_rs, log_offset,
  starting_betas, prop_burn = 0.1, return_cont = FALSE,
  beta_names = NA_character_, cont_names = NA_character_)
}
\arguments{
\item{counts}{a matrix of counts}

\item{design_mat}{design matrix for mean response}

\item{prior_sd_betas}{prior std. dev. for regression coefficients}

\item{rw_sd_betas}{random walk std. dev. for proposing beta values}

\item{prior_sd_rs}{prior std. dev for dispersion parameters}

\item{prior_mean_log_rs}{vector of prior means for dispersion parameters}

\item{n_it}{number of iterations to run MCMC}

\item{rw_sd_rs}{random walk std. dev. for proposing dispersion values}

\item{log_offset}{vector of offsets on log scale}

\item{design_mat_re}{design matrix for random intercepts}

\item{prior_sd_betas_a}{alpha in inverse gamma prior for random intercept variance}

\item{prior_sd_betas_b}{beta in inverse gamma prior for random intercept variance}
}
\value{
Returns a list with a cube of regression parameters, including random effects, a matrix of dispersion values, and a matrix of random intercept variances
}
\description{
Run an MCMC for the Negative Binomial mixed model (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
