% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbglm_mcmc_wls_gam}
\alias{nbglm_mcmc_wls_gam}
\title{Negative Binomial GLM MCMC WLS Gamma-disp (title)}
\usage{
nbglm_mcmc_wls_gam(counts, design_mat, prior_sd_betas, prior_shape, prior_scale,
  n_it, rw_sd_rs, log_offset, starting_betas, starting_disps, grain_size = 1L)
}
\arguments{
\item{counts}{a matrix of counts}

\item{design_mat}{designe matrix for mean response}

\item{prior_sd_betas}{prior std. dev. for regression coefficients}

\item{prior_shape}{vector of prior gamma shape parameters for dispersions}

\item{prior_scale}{vector of prior gamma scale parameters for dispersions}

\item{n_it}{number of iterations to run MCMC}

\item{rw_sd_rs}{random walk std. dev. for proposing dispersion values}

\item{log_offset}{vector of offsets on log scale}

\item{starting_betas}{matrix of starting values for regression coefficients (n_feature x n_beta)}

\item{starting_disps}{vector of starting values for dispersion parameters}

\item{grain_size}{minimum size of parallel jobs, defaults to 1, can ignore for now}
}
\value{
Returns a list with a cube of regression parameters and a matrix of dispersion values
}
\description{
Run an MCMC for the Negative Binomial (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
