% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbll_alpha_old}
\alias{nbll_alpha_old}
\title{Negative Binomial log-liklihood (reduced for dispersion)}
\usage{
nbll_alpha_old(counts, betas, alpha, design_mat)
}
\arguments{
\item{counts}{a vector of counts}

\item{betas}{a vector of regression coef}

\item{alpha}{multiplicative dispersion value}

\item{design_mat}{design matrix for mean response}
}
\value{
Returns a list with a cube of regression parameters, including random effects, a matrix of dispersion values, and a matrix of random intercept variances
}
\description{
Run an MCMC for the Negative Binomial mixed model (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
