% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbmm_mcmc_sampler_wls_force_fp}
\alias{nbmm_mcmc_sampler_wls_force_fp}
\title{Negative Binomial GLMM MCMC WLS Force (full parallel chians)}
\usage{
nbmm_mcmc_sampler_wls_force_fp(counts, design_mat, design_mat_re,
  prior_sd_betas, prior_sd_betas_a, prior_sd_betas_b, prior_sd_rs,
  prior_mean_log_rs, n_it, rw_sd_rs, log_offset, starting_betas,
  grain_size = 1L)
}
\arguments{
\item{counts}{a matrix of counts}

\item{design_mat}{design matrix for mean response}

\item{design_mat_re}{design matrix for random intercepts}

\item{prior_sd_betas}{prior std. dev. for regression coefficients}

\item{prior_sd_betas_a}{alpha in inverse gamma prior for random intercept variance}

\item{prior_sd_betas_b}{beta in inverse gamma prior for random intercept variance}

\item{prior_sd_rs}{prior std. dev for dispersion parameters}

\item{prior_mean_log_rs}{vector of prior means for dispersion parameters}

\item{n_it}{number of iterations to run MCMC}

\item{rw_sd_rs}{random wal std. dev. for proposing dispersion values}

\item{log_offset}{vector of offsets on log scale}

\item{grain_size}{minimum size of parallel jobs, defaults to 1, can ignore for now}
}
\value{
Returns a list with a cube of regression parameters, including random effects, a matrix of dispersion values, and a matrix of random intercept variances
}
\description{
Run an MCMC for the Negative Binomial mixed model (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
