% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbmm_mcmc_sampler_wls_force_fp_sum_cont_pb_nbf}
\alias{nbmm_mcmc_sampler_wls_force_fp_sum_cont_pb_nbf}
\title{Negative Binomial GLMM fit for RNA-Seq expression using MCMC (contrast version)}
\usage{
nbmm_mcmc_sampler_wls_force_fp_sum_cont_pb_nbf(counts, design_mat,
  design_mat_re, contrast_mat, prior_sd_betas, prior_sd_betas_a,
  prior_sd_betas_b, prior_sd_rs, prior_mean_log_rs, n_it, rw_sd_rs, log_offset,
  starting_betas, prop_burn_in = 0.1, grain_size = 1L, num_accept = 20L,
  abs_es = 0.05)
}
\arguments{
\item{counts}{A numeric matrix of RNA-Seq counts (rows are genes, columns are samples)}

\item{design_mat}{The fixed effects design matrix for mean response}

\item{design_mat_re}{The design matrix for random intercepts}

\item{contrast_mat}{A numeric matrix of linear contrasts of fixed effects to be tested.  Each row is considered to be an independent test, and each are done seperately}

\item{prior_sd_betas}{Prior std. dev. in normal prior for regression coefficients}

\item{prior_sd_betas_a}{Alpha parameter in inverse gamma prior for random intercept variance}

\item{prior_sd_betas_b}{Beta parameter in inverse gamma prior for random intercept variance}

\item{prior_sd_rs}{Prior std. dev in log-normal prior for dispersion parameters}

\item{prior_mean_log_rs}{Vector of prior means for log of dispersion parameters}

\item{n_it}{Number of iterations to run MCMC}

\item{rw_sd_rs}{Random walk std. dev. for proposing dispersion values (normal distribution centerd at current value)}

\item{log_offset}{Vector of offsets on log scale}

\item{starting_betas}{Numeric matrix of starting values for the regression coefficients.  For best results, supply starting values for at least the intercept (e.g. row means of counts matrix)}

\item{prop_burn_in}{Proportion of MCMC chain to discard as burn-in when computing summaries}

\item{grain_size}{minimum size of parallel jobs, defaults to 1, can ignore for now}

\item{num_accept}{Number of forced accepts of fixed and random effects at the beginning of the MCMC.  In practice forcing about 20 accepts (default value) prevents inverse errors at the start of chains and gives better mixing overall}
}
\value{
Returns a list with a cube of regression parameters, including random effects, a matrix of dispersion values, and a matrix of random intercept variances
}
\description{
Estimate Negative Binomial regressioin coefficients, dispersion parameter, and random intercept variance using MCMC with a weighted least squares proposal
}
\details{
This function estimates regression parameters and ...
}
\author{
Brian Vestal
}
