% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{whole_chain_nbglm_mala}
\alias{whole_chain_nbglm_mala}
\title{Single feature GLM chain using MALA proposal}
\usage{
whole_chain_nbglm_mala(counts, log_offset, starting_betas, design_mat, mean_rho,
  prior_sd_betas, prior_sd_rs, rw_sd_rs, n_beta, n_sample, n_it, h_step)
}
\arguments{
\item{counts}{a vector of counts}

\item{log_offset}{vector of offsets on log scale}

\item{starting_betas}{vector of starting values for betas}

\item{design_mat}{design matrix for mean response}

\item{mean_rho}{prior mean for dispersion parameter}

\item{prior_sd_betas}{prior std. dev. for regression coefficients}

\item{prior_sd_rs}{prior std. dev for dispersion parameters}

\item{rw_sd_rs}{random walk std. dev. for proposing dispersion values}

\item{n_beta}{random walk std. dev. for proposing dispersion values}

\item{n_sample}{random walk std. dev. for proposing dispersion values}

\item{n_it}{random walk std. dev. for proposing dispersion values}

\item{h_step}{random walk std. dev. for proposing dispersion values}
}
\value{
Returns a list with a cube of regression parameters, including random effects, a matrix of dispersion values, and a matrix of random intercept variances
}
\description{
Run an MCMC for the Negative Binomial mixed model (short description, one or two sentences)
}
\details{
This is where you write details on the function...

more details....
}
\author{
Brian Vestal
}
